+++ include/VEngine/Gfx/Resources/TextureManager.hpp
@@ -6,8 +6,6 @@
 
 #pragma once
 
-#include <memory>
-
 #include "VEngine/Gfx/Resources/Texture.hpp"
 
 namespace ven {
@@ -23,24 +21,24 @@ namespace ven {
             [[nodiscard]] static std::shared_ptr<Texture> getTexture(const Device& device, const SwapChain& swapChain, const std::string& filepath) {
                 auto& instance = getInstance();
                 if (instance.m_texturePaths.contains(filepath)) {
-                    const uint32_t index = instance.m_texturePaths[filepath];
+                    const uint8_t index = instance.m_texturePaths[filepath];
                     return instance.m_textures[index];
                 }
-                const uint32_t newIndex = instance.m_nextIndex++;
+                const uint8_t newIndex = instance.m_nextIndex++;
                 auto texture = std::make_shared<Texture>(device, swapChain, filepath);
                 instance.m_textures[newIndex] = texture;
                 instance.m_texturePaths[filepath] = newIndex;
                 return texture;
             }
 
-            [[nodiscard]] static const std::shared_ptr<Texture>& getTexturePath(const int index) {
+            [[nodiscard]] static const std::shared_ptr<Texture>& getTexturePath(const uint8_t index) {
                 if (auto& instance = getInstance(); instance.m_textures.contains(index)) {
                     return instance.m_textures[index];
                 }
                 throw utl::THROW_ERROR(("Texture not found with index: " + std::to_string(index)).c_str());
             }
 
-            [[nodiscard]] static uint32_t getTextureIndex(const std::string& filepath) {
+            [[nodiscard]] static uint8_t getTextureIndex(const std::string& filepath) {
                 if (auto& instance = getInstance(); instance.m_texturePaths.contains(filepath)) {
                     return instance.m_texturePaths[filepath];
                 }
@@ -53,14 +51,14 @@ namespace ven {
                 }
             }
 
-            [[nodiscard]] static const std::unordered_map<uint32_t, std::shared_ptr<Texture>>& getTextures() { return getInstance().m_textures; }
+            [[nodiscard]] static const std::unordered_map<uint8_t, std::shared_ptr<Texture>>& getTextures() { return getInstance().m_textures; }
             [[nodiscard]] static size_t getTextureSize() { return getInstance().m_textures.size(); }
 
         private:
 
-            uint32_t m_nextIndex = 0;
-            std::unordered_map<uint32_t, std::shared_ptr<Texture>> m_textures;
-            std::unordered_map<std::string, uint32_t> m_texturePaths;
+            uint8_t m_nextIndex = 0;
+            std::unordered_map<uint8_t, std::shared_ptr<Texture>> m_textures;
+            std::unordered_map<std::string, uint8_t> m_texturePaths;
 
             [[nodiscard]] static TextureManager& getInstance() {
                 static TextureManager instance;
